Add files via upload
https://commits.top/serbia.html
Most active GitHub users in Serbia
This is a list of most active GitHub users in Serbia over the past year. Other countries/cities can be found here. This list was generated at 2021-04-12 08:17:36.879903458 +0000 UTC m=+190.925293686.

This list is generated with this piece of code. The logic is as follows (in pseudocode):

githubUsers.sort(_.followers)
           .filter(_.location == 'Serbia') 
           .take(1000)
           .sort(_.contributions)
           .take(256)